+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.2.1] - 2025-08-01
 
 ### Fixed
+- Make 'resources' for the qbee_role resource optional.
+
 - Correctly handle drift detection for qbee_filemanager_file and qbee_filemanager_directory resources.
 
 ## [1.2.0] - 2025-06-04